+++ packages/models/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - Remove box.updateImage
 - Implement deletion of box images
+- Allow updating of box model value
 
 ## v0.0.9
 - Fix measurements with 'content-type: json' not parsed with JSON.parse.